style

+++ modules/matrix-calculator/src/matrix_calculator.cpp
@@ -99,19 +99,19 @@ MatrixCalculator MatrixCalculator::operator *(const MatrixCalculator& p) const {
 }
 
 double MatrixCalculator::Determinant() const {
-    if(matrix_.size() != matrix_[0].size()) {
+    if (matrix_.size() != matrix_[0].size()) {
         throw "Invalid sizes";
     }
 
     double tmp, det = 1;
     MatrixCalculator temp(*this);
 
-        for (int i = 0; i < temp.matrix_.size(); ++i) {
-            for (int j = i + 1; j < temp.matrix_.size(); ++j) {
+        for (size_t i = 0; i < temp.matrix_.size(); ++i) {
+            for (size_t j = i + 1; j < temp.matrix_.size(); ++j) {
                 if (matrix_[i][i] == 0)
                     throw "Division by zero!";
                 tmp = temp.matrix_[j][i] / temp.matrix_[i][i];
-                for (int k = i; k < temp.matrix_.size(); k++) {
+                for (size_t k = i; k < temp.matrix_.size(); k++) {
                     temp.matrix_[j][k] -= temp.matrix_[i][k] * tmp;
                 }
             }
